// components/student/studentMyProjectLEAN/studentMyProjectLEAN/studentMyProjectLEANListStyle.scss.d.ts
export type Styles = {
  'areaActionButtonsPreviewImageFile': string;
  'buttonAction': string;
  'buttonActionEdit': string;
  'buttonsActionsContainer': string;
  'buttonsActionsContainerNew': string;
  'imgContainer': string;
  'labelResourse': string;
  'simpleListItem': string;
};

export type ClassNames = keyof Styles;

declare const styles: Styles;

export default styles;